refactor: separate portable and cluster storage responsibilities
